Refactor UI into modular views
